refactor(pages): proof-read about, reuse experience counter
