+++ content/about.md
@@ -4,17 +4,17 @@ title: 'About'
 
 ### From idea to production
 
-The story of softbeehive began back in 2008 while I was studying mechanical engineering at National Aviation University in Kyiv. At the age of 19 I became disillusioned about my university and got bored by the slow learning process.
+The story of softbeehive began back in 2008 while I was studying mechanical engineering at National Aviation University in Kyiv, Ukraine. At the age of 19, I became disillusioned about my university and got bored by the slow learning process.
 
-I always wanted to jump straight into solving practical real-world issues, but I felt there was a little room for creativity. Jet engine in the lecture hall, I was so capivated by a year earlier, seemed like a museum exhibit.
+I always wanted to jump straight into solving practical, real-world issues, but felt there was a little room for creativity. The jet engine in the lecture hall, I was so captivated by a year earlier, seemed like a museum exhibit.
 
-Inspired by successful internet projects of that time, I decided to found a web studio. Few weeks later I completed first commercial project earning $100. That joy of reaching the goal shines in my collection of memories ever since.
+Inspired by successful internet projects of that period, I decided to found a web studio. A few weeks later, I completed the first commercial project earning $100. That joy of reaching the goal shines in my collection of memories ever since.
 
 ### Work philosophy
 
-In the course of *<experience-counter></experience-counter> years* I built many web apps. Society transformed, new tools and techniques evolved, flash died, but web browser is still a central app on our devices. And still due to the decentralized nature of the web, frontend devs have to deal with so many caveats.
+*Over <experience-counter></experience-counter> years*, I built many web apps. Society transformed, new tools and techniques evolved, javascript gained popularity, but web browser is still a central app on our devices. And yet, due to the web's decentralized nature, frontend devs have to deal with so many caveats.
 
-Modern responsive performant accessible website is a synonym to competitive edge. It doesn't matter which religion you practice: agile or lean. People values, their knowledge and working culture define the fate of business. Ability to focus on achieving the primary goal is what sets efficient teams apart.
+Modern, responsive, performant, accessible web app is a synonym to "competitive edge." It doesn't matter which religion you practice: agile or lean. People's values, knowledge, and working culture define the fate of the business. The ability to focus on achieving the primary goal is what sets efficient teams apart.
 
 ### Summary
-At work I care about look and feel, speed, security, accessibility and ethics of the project. I love the role at the intersection of software engineering and graphic design. This portfolio is a compressed snapshot of my personal and professional evolution.
+At work, I care about the look and feel, speed, security, accessibility, and ethics of the project. I love the role at the intersection of software engineering and graphic design. This portfolio is a compressed snapshot of my personal and professional evolution.
